#53 Extract Dogu relevant types into own files, update markdown generator
