+++ core/dogu_v2.go
@@ -1,7 +1,6 @@
 package core
 
 import (
-	"encoding/json"
 	"errors"
 	"fmt"
 	"maps"
@@ -44,131 +43,6 @@ const (
 	// Deprecated: This field is no longer used. There is no substitute.
 	ExposedCommandPostBackup = "post-backup"
 )
-
-// VolumeClient adds additional information for clients to create volumes.
-//
-// Example:
-//
-//	{
-//	  "Name": "k8s-dogu-operator",
-//	  "Params": {
-//	    "Type": "configmap",
-//	    "Content": {
-//	      "Name": "k8s-ces-menu-json"
-//	    }
-//	  }
-//	}
-type VolumeClient struct {
-	// Name identifies the client responsible to process this volume definition. This field is mandatory.
-	//
-	// Examples:
-	//   - cesapp
-	//   - k8s-dogu-operator
-	Name string
-	// Params contains generic data only interpretable by the client. This field is mandatory.
-	Params interface{}
-}
-
-// Volume defines container volumes that are created during the dogu creation or upgrade.
-//
-// Examples:
-//   - { "Name": "data", "Path":"/usr/share/yourtool/data", "Owner":"1000", "Group":"1000", "NeedsBackup": true}
-//   - { "Name": "temp", "Path":"/tmp", "Owner":"1000", "Group":"1000", "NeedsBackup": false}
-type Volume struct {
-	// Name identifies the volume. This field is mandatory. It must be unique in all volumes of the same dogu.
-	//
-	// The name syntax must comply with the file system syntax of the respective host operating system and is encouraged
-	// to consist of:
-	//   - lower case latin characters
-	//   - special characters underscore "_", minus "-"
-	//   - ciphers 0-9
-	//
-	// The name must not be "_private" to avoid conflicts with the dogu's private key.
-	//
-	// Examples:
-	//   - tooldata
-	//   - tool-data-0
-	//
-	Name string
-	// Path to the directory or file where the volume will be mounted inside the dogu. This field is mandatory.
-	//
-	// Path may consist of several directory levels, delimited by a forward slash "/". Path must comply with the file
-	// system syntax of the container operating system.
-	//
-	// The path must not match `/private` to avoid conflicts with the dogu's private key.
-	//
-	// Examples:
-	//   - /usr/share/yourtool
-	//   - /tmp
-	//   - /usr/share/license.txt
-	//
-	Path string
-	// Owner contains the numeric Unix UID of the user owning this volume. This field is optional.
-	//
-	// For security reasons it is strongly recommended to set the Owner of the volume to an unprivileged user. Please
-	// note that container image must be then built in a way that the container process may own the path either by
-	// user or group ownership.
-	//
-	// The owner syntax must consist of ciphers (0-9) only.
-	//
-	// Examples:
-	//   - "1000" - an unprivileged user
-	//   - "0" - the root user
-	//
-	Owner string
-	// Group contains the numeric Unix GID of the group owning this volume. This field is optional.
-	//
-	// For security reasons it is strongly recommended to set the Group of the volume to an unprivileged Group. Please
-	// note that container image must be then built in a way that the container process may own the path either by
-	// user or group ownership.
-	//
-	// The Group syntax must consist of ciphers (0-9) only.
-	//
-	// Examples:
-	//   - "1000" - an unprivileged group
-	//   - "0" - the root group
-	//
-	Group string
-	// NeedsBackup controls whether the Cloudogu EcoSystem backup facility backs up the whole the volume or not. This
-	// field is optional. If unset, a value of `false` will be assumed.
-	NeedsBackup bool
-	// Clients contains a list of client-specific (t. i., the client that interprets the dogu.json) configurations for
-	// the volume. This field is optional.
-	//
-	Clients []VolumeClient `json:"Clients,omitempty"`
-}
-
-// GetClient retrieves a client with a given name and return a pointer to it. If a client does not exist a nil pointer
-// and false are returned.
-func (v *Volume) GetClient(clientName string) (*VolumeClient, bool) {
-	if v.Clients == nil {
-		return nil, false
-	}
-
-	for i := range v.Clients {
-		if v.Clients[i].Name == clientName {
-			return &v.Clients[i], true
-		}
-	}
-
-	return nil, false
-}
-
-// UnmarshalJSON sets the default value for NeedsBackup. We are preventing an infinite loop by using a local Alias type
-// to call json.Unmarshal again
-func (v *Volume) UnmarshalJSON(data []byte) error {
-	type Alias Volume
-	a := &struct {
-		*Alias
-	}{
-		Alias: (*Alias)(v),
-	}
-	v.NeedsBackup = true
-	if err := json.Unmarshal(data, &a); err != nil {
-		return err
-	}
-	return nil
-}
 
 // HealthCheck provide readiness and health checks for the dogu container.
 type HealthCheck struct {
@@ -319,345 +193,11 @@ type ServiceAccount struct {
 	Kind string `json:"Kind,omitempty"`
 }
 
-// ConfigurationField describes a single dogu configuration field which is stored in the Cloudogu EcoSystem registry.
-type ConfigurationField struct {
-	// Name contains the name of the configuration key. The field is mandatory. It
-	// must not contain leading or trailing slashes "/", but it may contain
-	// directory keys delimited with slashes "/" within the name.
-	//
-	// The Name syntax is encouraged to consist of:
-	//   - lower case latin characters
-	//   - special characters underscore "_"
-	//   - ciphers 0-9
-	//
-	// Example:
-	//   - feedback_url
-	//   - logging/root
-	//
-	Name string
-	// Description offers context and purpose of the configuration field in human-readable format. This field is
-	// optional, yet highly recommended to be set.
-	//
-	// Example:
-	//   - "Set the root log level to one of ERROR, WARN, INFO, DEBUG or TRACE. Default is INFO"
-	//   - "URL of the feedback service"
-	//
-	Description string
-	// Optional allows to have this configuration field unset, otherwise a value must be set. This field is optional.
-	// If unset, a value of `false` will be assumed.
-	//
-	// Example:
-	//   - true
-	//
-	Optional bool
-	// Encrypted marks this configuration field to contain a sensitive value that will be encrypted with the dogu's
-	// public key. This field is optional. If unset, a value of `false` will be assumed.
-	//
-	// Example:
-	//   - true
-	//
-	Encrypted bool
-	// Global marks this configuration field to contain a value that is available for all dogus. This field is optional.
-	// If unset, a value of `false` will be assumed.
-	//
-	// Example:
-	//   - true
-	//
-	Global bool
-	// Default defines a default value that may be evaluated if no value was configured, or the value is empty or even
-	// invalid. This field is optional.
-	//
-	// Example:
-	//   - "WARN"
-	//   - "true"
-	//   - "https://scm-manager.org/plugins"
-	//
-	Default string
-	// Validation configures a validator that will be used to mark invalid or
-	// out-of-range values for this configuration field. This field is optional.
-	//
-	// Example:
-	//  "Validation": {
-	//     "Type": "ONE_OF", // only allows one of these two values
-	//     "Values": [
-	//       "value 1",
-	//       "value 2",
-	//     ]
-	//   }
-	//
-	//  "Validation": {
-	//     "Type": "FLOAT_PERCENTAGE_HUNDRED" // valid values range between 0.0 and 100.0
-	//  }
-	//
-	//  "Validation": {
-	//    "Type": "BINARY_MEASUREMENT" // only allows suffixed integer values measured in byte, kibibyte, mebibyte, gibibyte
-	//   }
-	//
-	//
-	Validation ValidationDescriptor
-}
-
-// ValidationDescriptor describes how to determine if a config value is valid.
-type ValidationDescriptor struct {
-	// Type contains the name of the config value validator. This field is mandatory. Valid types are:
-	//
-	//   - ONE_OF
-	//   - BINARY_MEASUREMENT
-	//   - FLOAT_PERCENTAGE_HUNDRED
-	//
-	Type string
-	// Values may contain values that aid the selected validator. The values may or
-	// may not be optional, depending on the Type being used.
-	// It is up to the selected validator whether this field is mandatory, optional,
-	// or unused.
-	Values []string
-}
-
 // Properties describes generic properties of the dogu which are evaluated by a client like cesapp or k8s-dogu-operator.
 //
 // Example:
 //   - { "key": "value" }
 type Properties map[string]string
-
-// Contains the different kind of types supported by dogu dependencies
-const (
-	// DependencyTypeDogu identifies a dogu dependency towards another dogu.
-	DependencyTypeDogu = "dogu"
-	// DependencyTypeClient identifies a dogu dependency towards a dogu.json-processing client like cesapp or
-	// k8s-dogu-operator.
-	DependencyTypeClient = "client"
-	// DependencyTypePackage identifies a dogu dependency towards an operating system package.
-	DependencyTypePackage = "package"
-)
-
-// Dependency describes the quality of a dogu dependency towards another entity.
-//
-// Examples:
-//
-//	{
-//	  "type": "dogu",
-//	  "name": "postgresql"
-//	}
-//
-//	{
-//	  "name": "postgresql"
-//	}
-//
-//	{
-//	  "type": "client",
-//	  "name": "k8s-dogu-operator",
-//	  "version": ">=0.16.0"
-//	}
-//
-//	{
-//	  "type": "package",
-//	  "name": "cesappd",
-//	  "version": ">=3.2.0"
-//	}
-type Dependency struct {
-	// Type identifies the entity on which the dogu depends. This field is optional.
-	// If unset, a value of `dogu` is then assumed.
-	//
-	// Valid values are one of these: "dogu", "client", "package".
-	//
-	// A type of "dogu" references another dogu which must be present and running
-	// during the dependency check.
-	//
-	// A type of "client" references the client which processes this dogu's
-	// "dogu.json". Several client dependencies of a different client type can be
-	// used f. i. to prohibit the processing of a certain client.
-	//
-	// A type of "package" references a necessary operating system package that must
-	// be present during the dependency check.
-	//
-	// Examples:
-	//  - "dogu"
-	//  - "client"
-	//  - "package"
-	//
-	Type string `json:"type"`
-	// Name identifies the entity selected by Type. This field is mandatory. If the Type selects another dogu, Name
-	// must use the simple dogu name (f. e. "postgres"), not the full qualified dogu name (not "official/postgres").
-	//
-	// Examples:
-	//  - "postgresql"
-	//  - "k8s-dogu-operator"
-	//  - "cesappd"
-	//
-	Name string `json:"name"`
-	// Version selects the version of entity selected by Type. This field is optional. If unset, any version of the
-	// selected entity will be accepted during the dependency check.
-	//
-	// Version accepts different version styles and compare operators.
-	//
-	// Examples:
-	//
-	//  - ">=4.1.1-2" - select the entity version greater than or equal to version 4.1.1-2
-	//  - "<=1.0.1" - select the entity version less than or equal to version 1.0.1
-	//  - "1.2.3.4" - select exactly the version 1.2.3.4
-	//
-	// With a non-existing version it is possible to negate a dependency.
-	//
-	// Example:
-	//
-	//   - "<=0.0.0" - prohibit the selected entity being present
-	//
-	Version string `json:"version"`
-}
-
-// Capability represent POSIX capabilities type.
-//
-// See docs at https://manned.org/capabilities.7
-type Capability string
-
-// DefaultCapabilities in the Cloudogu Ecosystem, regardless of which container runtime it runs in.
-// This way the Ecosystem can guarantee a consistent behavior.
-var DefaultCapabilities = []Capability{
-	Chown, DacOverride, Fsetid, Fowner, Setgid, Setuid, Setpcap, NetBindService, Kill,
-}
-
-var allCapabilities = []Capability{AuditControl, AuditRead, AuditWrite, BlockSuspend, Bpf, CheckpointRestore, Chown, DacOverride, Fowner, Fsetid, IpcLock, IpcOwner, Kill, Lease, LinuxImmutable, MacAdmin, MacOverride, Mknod, NetAdmin, NetBindService, NetBroadcast, NetRaw, Perfmon, Setfcap, Setgid, Setpcap, Setuid, SysAdmin, SysBoot, SysChroot, SysModule, SysNice, SysPAcct, SysPTrace, SysResource, SysTime, SysTtyCONFIG, Syslog, WakeAlarm}
-
-// These capabilities' documentation contain abstracts of their respective manpage documentation and may refer to
-// other man pages references f. e. as epoll(7)
-const (
-	// All is a special capability which can be used to add or drop all capabilities listed below.
-	All = "ALL"
-	// AuditControl enables and disables kernel auditing; changes auditing filter rules.
-	// retrieves auditing status and filtering rules.
-	AuditControl = "AUDIT_CONTROL"
-	// AuditRead allows reading the audit log via a multicast netlink socket.
-	AuditRead = "AUDIT_READ"
-	// AuditWrite write records to kernel auditing log.
-	AuditWrite = "AUDIT_WRITE"
-	// BlockSuspend employs features that can block system suspend (epoll(7) EPOLLWAKEUP, /proc/sys/wake_lock).
-	BlockSuspend = "BLOCK_SUSPEND"
-	// Bpf employs privileged BPF operations and separates out BPF functionality from the overloaded SYS_ADMIN capability.
-	Bpf = "BPF"
-	// CheckpointRestore allows facilitating checkpoint/restore for non-root users.
-	CheckpointRestore = "CHECKPOINT_RESTORE"
-	// Chown makes arbitrary changes to file UIDs and GIDs (see chown(2)).
-	Chown = "CHOWN"
-	// DacOverride allows bypassing file read, write, and execute permission checks to implement discretionary access control.
-	DacOverride = "DAC_OVERRIDE"
-	// Fowner allows to bypass permission checks on operations that require file ownership (i. e. read, write, or execute).
-	Fowner = "FOWNER"
-	// Fsetid allows to avoid resetting a file's Setuid flag on modifying.
-	Fsetid = "FSETID"
-	// IpcLock allows a process to lock memory to swapping the content to disk.
-	IpcLock = "IPC_LOCK"
-	// IpcOwner bypasses permission checks for operations on System V IPC objects.
-	IpcOwner = "IPC_OWNER"
-	// Kill bypasses permission checks for sending signals (see kill(2)). This includes use of the ioctl(2) KDSIGACCEPT operation.
-	Kill = "KILL"
-	// Lease establishes leases on arbitrary files (see fcntl(2)).
-	Lease = "LEASE"
-	// LinuxImmutable Set the FS_APPEND_FL and FS_IMMUTABLE_FL inode flags (see FS_IOC_SETFLAGS(2const)).
-	LinuxImmutable = "LINUX_IMMUTABLE"
-	// MacAdmin allows MAC configuration or state changes. Implemented for the Smack Linux Security Module (LSM).
-	MacAdmin = "MAC_ADMIN"
-	// MacOverride Override Mandatory Access Control (MAC).  Implemented for the Smack LSM.
-	MacOverride = "MAC_OVERRIDE"
-	// Mknod creates special files using mknod(2).
-	Mknod = "MKNOD"
-	// NetAdmin allows processes to perform various network-related operations.
-	NetAdmin = "NET_ADMIN"
-	// NetBindService binds a socket to Internet domain privileged ports (port numbers less than 1024).
-	NetBindService = "NET_BIND_SERVICE"
-	// NetBroadcast (Unused) makes socket broadcasts, and listen to multicasts.
-	NetBroadcast = "NET_BROADCAST"
-	// NetRaw uses RAW and PACKET sockets; binds to any address for transparent proxying.
-	NetRaw = "NET_RAW"
-	// Perfmon employs various performance-monitoring mechanisms various BPF operations that have performance implications.
-	Perfmon = "PERFMON"
-	// Setfcap allows to set arbitrary capabilities on a file. Since Linux 5.12, this capability is also needed to map
-	// user ID 0 in a new user namespace; see user_namespaces(7) for details.
-	Setfcap = "SETFCAP"
-	// Setgid makes arbitrary manipulations of process GIDs and supplementary GID list;
-	//    - forge GID when passing socket credentials via UNIX domain sockets;
-	//    - write a group ID mapping in a user namespace (see user_namespaces(7)).
-	Setgid = "SETGID"
-	// Setpcap allows to add or drop any capability from the calling thread's bounding set.
-	Setpcap = "SETPCAP"
-	// Setuid allows to make arbitrary manipulations of process UIDs (setuid(2), setreuid(2), setresuid(2), setfsuid(2))
-	Setuid = "SETUID"
-	//	SysAdmin allows performing a range of system administration operations. This is like super dangerous and should not be granted easily
-	SysAdmin = "SYS_ADMIN"
-	//SysBoot allows to use reboot(2) and kexec_load(2).
-	SysBoot = "SYS_BOOT"
-	// SysChroot allows to use chroot(2) or change mount linux namespaces using setns(2).
-	SysChroot = "SYS_CHROOT"
-	// SysModule loads and unloads kernel modules (see init_module(2) and delete_module(2))
-	SysModule = "SYS_MODULE"
-	// SysNice lowers the process nice value (nice(2), setpriority(2)) and change the nice value for arbitrary processes
-	SysNice = "SYS_NICE"
-	//SysPAcct switches process accounting with acct(2) on or off.
-	SysPAcct = "SYS_PACCT"
-	//SysPTrace allows to trace arbitrary processes using ptrace(2), or read/write memory of arbitrary processes.
-	SysPTrace = "SYS_PTRACE"
-	// SysResource allows a process to modify the resource limits specified in variety of uses cases.
-	SysResource = "SYS_RESOURCE"
-	// SysTime sets system clock (settimeofday(2), stime(2), adjtimex(2)); set real-time (hardware) clock
-	SysTime = "SYS_TIME"
-	// SysTtyCONFIG uses vhangup(2); employ various privileged ioctl(2) operations on virtual terminals.
-	SysTtyCONFIG = "SYS_TTY_CONFIG"
-	// Syslog performs privileged syslog(2) operations.
-	Syslog = "SYSLOG"
-	// WakeAlarm triggers something that will wake up the system.
-	WakeAlarm = "WAKE_ALARM"
-)
-
-// Capabilities represent POSIX capabilities that can be added to or removed from a dogu.
-//
-// The fields Add and Drop will modify the default capabilities as provided by DefaultCapabilities. Add will append
-// further capabilities while Drop will remove capabilities. The capability All can be used to add or remove all
-// available capabilities.
-//
-// See DefaultCapabilities for the standard set being used in the Cloudogu Ecosystem.
-//
-// This example will result in the following capability list: DacOverride, Fsetid, Fowner, Setgid, Setuid, Setpcap, NetBindService, Kill, Syslog
-//
-//	"Capabilities": {
-//	   "Drop": "Chown"
-//	   "Add": "Syslog"
-//	}
-//
-// This example will result in the following capability list: NetBindService
-//
-//	"Capabilities": {
-//	   "Drop": ["All"],
-//	   "Add": ["NetBindService", "Kill"]
-//	}
-type Capabilities struct {
-	// Add contains the capabilities that should be allowed to be used in a container. This list is optional.
-	Add []Capability `json:"Add,omitempty"`
-	// Drop contains the capabilities that should be blocked from being used in a container. This list is optional.
-	Drop []Capability `json:"Drop,omitempty"`
-}
-
-// Security defines security policies for the dogu. These fields can be used to reduce a dogu's attack surface.
-//
-// Example:
-//
-//	"Security": {
-//	  "Capabilities": {
-//	     "Drop": ["All"],
-//	     "Add": ["NetBindService", "Kill"]
-//	   },
-//	  "RunAsNonRoot": true,
-//	  "ReadOnlyRootFileSystem": true
-//	}
-type Security struct {
-	// Capabilities sets the allowed and dropped capabilities for the dogu. The dogu should not use more than the
-	// configured capabilities here, otherwise failure may occur at start-up or at run-time. This list is optional.
-	Capabilities Capabilities `json:"Capabilities,omitempty"`
-	// RunAsNonRoot indicates that the container must run as a non-root user. The dogu must support running as non-root
-	// user otherwise the dogu start may fail. This flag is optional and defaults to false.
-	RunAsNonRoot bool
-	// ReadOnlyRootFileSystem mounts the container's root filesystem as read-only. The dogu must support accessing the
-	// root file system by only reading otherwise the dogu start may fail. This flag is optional and defaults to false.
-	ReadOnlyRootFileSystem bool
-}
 
 // Dogu describes properties of a containerized application for the Cloudogu EcoSystem. Besides the meta information and
 // the [OCI container image], Dogu describes all necessities for automatic container instantiation, f. i. volumes,
@@ -1457,91 +997,4 @@ func (d *Dogu) ValidateSecurity() error {
 	}
 
 	return nil
-}
-
-// DoguJsonV2FormatProvider provides methods to format Dogu results compatible to v2 API.
-type DoguJsonV2FormatProvider struct{}
-
-// GetVersion returns DoguApiV2 for this implementation.
-func (d *DoguJsonV2FormatProvider) GetVersion() DoguApiVersion {
-	return DoguApiV2
-}
-
-// ReadDoguFromString reads a dogu from a string and returns the API v2 representation.
-func (d *DoguJsonV2FormatProvider) ReadDoguFromString(content string) (*Dogu, error) {
-	var dogu *Dogu
-	err := json.Unmarshal([]byte(content), &dogu)
-	if err != nil {
-		return nil, err
-	}
-
-	return dogu, nil
-}
-
-// ReadDogusFromString reads multiple dogus from a string and returns the API v2 representation.
-func (d *DoguJsonV2FormatProvider) ReadDogusFromString(content string) ([]*Dogu, error) {
-	var dogus []*Dogu
-	err := json.Unmarshal([]byte(content), &dogus)
-	if err != nil {
-		return nil, err
-	}
-
-	return dogus, nil
-}
-
-// WriteDoguToString receives a single dogu and returns the API v2 representation.
-func (d *DoguJsonV2FormatProvider) WriteDoguToString(dogu *Dogu) (string, error) {
-	data, err := json.Marshal(dogu)
-	if err != nil {
-		return "", err
-	}
-	return string(data), err
-}
-
-// WriteDogusToString receives a list of dogus and returns the API v2 representation.
-func (d *DoguJsonV2FormatProvider) WriteDogusToString(dogu []*Dogu) (string, error) {
-	data, err := json.Marshal(dogu)
-	if err != nil {
-		return "", err
-	}
-	return string(data), err
-}
-
-// ByDoguVersion implements sort.Interface for []Dogu to Dogus by their versions
-type ByDoguVersion []*Dogu
-
-// Len is the number of elements in the collection.
-func (doguVersions ByDoguVersion) Len() int {
-	return len(doguVersions)
-}
-
-// Swap swaps the elements with indexes i and j.
-func (doguVersions ByDoguVersion) Swap(i, j int) {
-	doguVersions[i], doguVersions[j] = doguVersions[j], doguVersions[i]
-}
-
-// Less reports whether the element with index i should sort before the element with index j.
-func (doguVersions ByDoguVersion) Less(i, j int) bool {
-	v1, err := ParseVersion(doguVersions[i].Version)
-	if err != nil {
-		GetLogger().Errorf("connot parse version %s for comparison", doguVersions[i].Version)
-	}
-	v2, err := ParseVersion(doguVersions[j].Version)
-	if err != nil {
-		GetLogger().Errorf("connot parse version %s for comparison", doguVersions[j].Version)
-	}
-
-	isNewer := v1.IsNewerThan(v2)
-	return isNewer
-}
-
-// ContainsDoguWithName checks if a dogu is contained in a slice by comparing the full name (including namespace)
-func ContainsDoguWithName(dogus []*Dogu, name string) bool {
-	for _, dogu := range dogus {
-		if dogu.Name == name {
-			return true
-		}
-	}
-
-	return false
 }
